F function adaptation for polygons

test statistic for the MC test.

pf12(obj_sp1, obj_sp2, r_min = 1, r_max = NULL, by = 1)

Arguments

obj_sp1
a SpatialPolygons or SpatialPointsDataFrame object

obj_sp2
a SpatialPolygons or SpatialPointsDataFrame object

r_min
a numeric representing the mininmun distance to calculate the function, default = 0.

r_max
a numeric representing the maximum distance to calculate the function default = max(polygonds distance matrix).

by
a numeric value that represents the how many units between each r value.

Value

a numeric matrix corresponding to the estimated \(F_{1,2}\) and \(F_{2,1}\) function in the interval \([r_{min}, r_{max}]\).